+++ frontend/src/components/TransactionDetails.js
@@ -13,6 +13,8 @@ const TransactionDetails = ({ transaction }) => {
   const { language } = useLangContext();
   
   const [lang, setLang] = useState(English);
+  const [trashClicks, setTrashClicks] = useState(0);
+  const [trashIconClass, setTrashIconClass] = useState('material-symbols-outlined trash-icon-default')
   const [usedDate, setUsedDate] = useState(formatEnglishDate(generatedDate));
 
   useEffect(() => {
@@ -26,23 +28,54 @@ const TransactionDetails = ({ transaction }) => {
       setUsedDate(formatSpanishDate(generatedDate));
     }
   }, [language, generatedDate])
+
+  useEffect(() => {
+    document.body.addEventListener('click', removeFocusedTrashIcon )
+
+    return function cleanup() {
+      window.removeEventListener('click', removeFocusedTrashIcon )
+    }
+  }, [trashClicks])
+
+  const removeFocusedTrashIcon = (e) => {
+    
+    if (e.target.innerHTML === 'delete') {
+      return
+    }
+
+    if (trashClicks > 0) {
+      setTrashClicks(0)
+      setTrashIconClass('material-symbols-outlined trash-icon-default')
+    }
+    
+  }
   
-  const handleClick = async () => {
+  const handleClick = async (e) => {
     if (!user) {
       return
     }
 
-    const res = await fetch('https://money-dunn-right.onrender.com/api/expenses/' + transaction._id, {
-      method: 'DELETE',
-      headers: {
-        'Authorization': `Bearer ${user.token}`
-      }
-    });
-    const json = await res.json();
-
-    if (res.ok) {
-      dispatch({type: 'DELETE_TRANSACTION', payload: json})
+    if (trashClicks == 0) {
+      setTrashClicks(1)
+      setTrashIconClass('material-symbols-outlined trash-icon-one-click')
+      return
     }
+
+    if (trashClicks > 0) {
+      const res = await fetch('https://money-dunn-right.onrender.com/api/expenses/' + transaction._id, {
+        method: 'DELETE',
+        headers: {
+          'Authorization': `Bearer ${user.token}`
+        }
+      });
+      const json = await res.json();
+
+      if (res.ok) {
+        dispatch({type: 'DELETE_TRANSACTION', payload: json})
+      }
+    }
+
+    
   }
 
   return (
@@ -54,7 +87,7 @@ const TransactionDetails = ({ transaction }) => {
       <p>{usedDate}</p>
       <span
        onClick={handleClick}
-       className='material-symbols-outlined'
+       className={trashIconClass}
       >delete
       </span>
     </div>
